+++ changelogs/mediation/applovin/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to AppLovin Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 12.6.0.3
+
+#### Updated
+* Yandex Mobile Ads SDK version 7.6.0
+
 ## Version 12.6.0.2
 
 #### Updated