+++ changelogs/mediation/google/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to AdMob Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 23.4.0.0
+
+#### Updated
+* Yandex Mobile Ads SDK version 7.6.0
+* Google SDK version 23.4.0
+
 ## Version 23.2.0.0
 
 #### Updated